+++ classes/Generator/Generator/Type.php
@@ -45,6 +45,12 @@ class Generator_Generator_Type
 	protected $_file;
 
 	/**
+	 * The base path in which the file should be created
+	 * @var string
+	 */
+	protected $_path;
+
+	/**
 	 * The base folder in which the file should be created
 	 * @var string
 	 */
@@ -157,6 +163,37 @@ class Generator_Generator_Type
 	}
 
 	/**
+	 * Setter and getter for the absolute base path in which the generator items will
+	 * be created. If not set, either APPPATH or MODPATH will be used by default.
+	 *
+	 * @param   string  $path  The absolute base file path
+	 * @return  string|Generator_Type  The current base path or this instance
+	 * @throws  Generator_Exception  On missing base path
+	 */
+	public function path($path = NULL)
+	{
+		if ($path === NULL)
+		{
+			if ( ! $this->_path)
+				return $this->_module ? MODPATH : APPPATH;
+
+			if ($this->_verify AND ! file_exists($this->_path))
+				throw new Generator_Exception("Path ':path' does not exist", array(':path' => $this->_path));
+
+			return $this->_path;
+		}
+
+		if ($path AND substr($path, -1, 1) != DIRECTORY_SEPARATOR)
+		{
+			$path .= DIRECTORY_SEPARATOR;
+		}
+
+		$this->_path = $path;
+
+		return $this;
+	}
+
+	/**
 	 * Setter and getter for the base folder in which the generator items will
 	 * be created, e.g. 'classes', 'tests'.
 	 *
@@ -176,7 +213,7 @@ class Generator_Generator_Type
 	/**
 	 * Setter and getter for the module in which generator items will be created.
 	 * This must be either the name of a loaded module as defined in the bootstrap,
-	 * or a valid folder under the current MODPATH.
+	 * or a valid folder under the current MODPATH or defined custom base path.
 	 *
 	 * @param   string  $module  The module name
 	 * @return  string|Generator_Type  The current module name or this instance
@@ -370,7 +407,7 @@ class Generator_Generator_Type
 
 		// Determine the base path for the file
 		$path = $this->_module ? Generator::get_module_path($this->_module,
-			$this->_verify) : APPPATH;
+			$this->_verify, $this->path()) : $this->path();
 
 		// Get the file name, optionally converting it to a path
 		$name = $convert ? (str_replace('_', $ds, $this->_name)) : $this->_name;
@@ -663,11 +700,10 @@ class Generator_Generator_Type
 
 		// Set the base path
 		$base = rtrim(($this->_module ? dirname(Generator::get_module_path(
-			$this->_module, $this->_verify)) : APPPATH), $ds);
+			$this->_module, $this->_verify, $this->path())) : $this->path()), $ds);
 
 		// Remove each leaf from the path
-		while (($path = substr($path, 0, strrpos($path, $ds)))
-			AND $path != $base)
+		while (($path = substr($path, 0, strrpos($path, $ds))) AND $path != $base)
 		{
 			// Add the sub-path to the list
 			array_unshift($tree, $path);